+++ src/suggestion/index.ts
@@ -1,7 +1,12 @@
 import type { App } from 'vue';
 import Suggestion from './Suggestion.vue';
 
-export type { SuggestionProps, SuggestionSemanticType, SuggestionItem } from './interface';
+export type {
+  SuggestionProps,
+  SuggestionSemanticType,
+  SuggestionItem,
+  SuggestionCascaderPassthrough,
+} from './interface';
 
 // @ts-ignore
 Suggestion.install = function(app: App) {